+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/model/order/CreateOrderRequest.kt
@@ -1,7 +1,5 @@
 package org.wordpress.android.fluxc.model.order
 
-import org.wordpress.android.fluxc.model.WCOrderModel.LineItem
-
 data class CreateOrderRequest(
     val lineItems: List<LineItem>,
     val shippingAddress: OrderAddress.Shipping,
